채널 구독 기능과 알림 기능 추가, 알림은 추후 분리도 생각

+++ src/main/java/com/live/main/user/controller/UserController.java
@@ -3,7 +3,9 @@ package com.live.main.user.controller;
 import com.live.main.channel.database.dto.ChannelDto;
 import com.live.main.channel.service.Interface.ChannelServiceInterface;
 import com.live.main.channel.service.Interface.CoverServiceInterface;
+import com.live.main.channel.service.Interface.PostServiceInterface;
 import com.live.main.common.database.dto.ErrorCode;
+import com.live.main.common.database.repository.OnlineRepository;
 import com.live.main.common.exception.CustomException;
 import com.live.main.profile.service.Interface.ProfileServiceInterface;
 import com.live.main.user.database.dto.CustomUserDetails;
@@ -43,7 +45,10 @@ public class UserController {
   private final VideoServiceInterface videoService;
   private final ProfileServiceInterface profileService;
   private final CoverServiceInterface coverService;
+  private final PostServiceInterface postService;
   private final JwtService jwtService;
+
+  private final OnlineRepository onlineRepository;
 
   @PostMapping("/register")
   public ResponseEntity<?> UserRegister(@RequestBody UserDto userDto){
@@ -89,6 +94,7 @@ public class UserController {
 
     response.addHeader(HttpHeaders.SET_COOKIE,newAccessCookie.toString());
     response.addHeader(HttpHeaders.SET_COOKIE, newRefreshCookie.toString());
+
     result.put("result", true);
 
     return ResponseEntity.ok(result);
@@ -111,7 +117,7 @@ public class UserController {
 
     response.addHeader(HttpHeaders.SET_COOKIE,clearAccessCookie.toString());
     response.addHeader(HttpHeaders.SET_COOKIE, clearRefreshCookie.toString());
-
+    onlineRepository.delete(principal.getUserid());
     return ResponseEntity.ok().build();
   }
 
@@ -190,9 +196,14 @@ public class UserController {
     boolean cover_delete= coverService.cover_delete_on_channel(channelDto.getName());
     boolean channel_delete = channelService.deleteChannelOnUser(user.getLoginId());
     boolean profile_delete= profileService.profile_delete_onUser(user.getLoginId());
+    boolean post_delete= postService.deletePostOnChannel(channelDto.getName());
+    boolean subscription_delete_1= channelService.deleteSubscriptionOnChannel(channelDto.getName());
+    boolean subscription_delete_2= channelService.deleteSubscriptionOnUser(user.getLoginId());
     boolean delete=userService.DeleteUser(user);
 
-    if(video_delete && cover_delete && channel_delete && profile_delete && delete){
+    if(video_delete && cover_delete && channel_delete
+       && profile_delete && post_delete &&subscription_delete_1
+       && subscription_delete_2 && delete){
       jwtService.deleteRefreshToken(user.getLoginId());
       ResponseCookie clearAccessCookie=ResponseCookie.from("accessToken","")
           .httpOnly(true).secure(true).sameSite("None").path("/").maxAge(0)
